+++ packages/services/protobuf/go/ecs-snapshot/ecs-snapshot_grpc.pb.go
@@ -26,8 +26,16 @@ type ECSStateSnapshotServiceClient interface {
 	GetStateLatest(ctx context.Context, in *ECSStateRequestLatest, opts ...grpc.CallOption) (*ECSStateReply, error)
 	// Requests the latest ECS state in stream format, which will chunk the state.
 	GetStateLatestStream(ctx context.Context, in *ECSStateRequestLatestStream, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamClient, error)
+	// Requests the latest ECS state in stream format, which will chunk the state.
+	//
+	// V2 version optimized to return entities as raw bytes.
+	GetStateLatestStreamV2(ctx context.Context, in *ECSStateRequestLatestStream, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamV2Client, error)
 	// Requests the latest ECS state, with aditional pruning.
 	GetStateLatestStreamPruned(ctx context.Context, in *ECSStateRequestLatestStreamPruned, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamPrunedClient, error)
+	// Requests the latest ECS state, with aditional pruning.
+	//
+	// V2 version optimized to return entities as raw bytes.
+	GetStateLatestStreamPrunedV2(ctx context.Context, in *ECSStateRequestLatestStreamPruned, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamPrunedV2Client, error)
 	// Requests the latest block number based on the latest ECS state.
 	GetStateBlockLatest(ctx context.Context, in *ECSStateBlockRequestLatest, opts ...grpc.CallOption) (*ECSStateBlockReply, error)
 	// Requests the ECS state at specific block.
@@ -85,8 +93,40 @@ func (x *eCSStateSnapshotServiceGetStateLatestStreamClient) Recv() (*ECSStateRep
 	return m, nil
 }
 
+func (c *eCSStateSnapshotServiceClient) GetStateLatestStreamV2(ctx context.Context, in *ECSStateRequestLatestStream, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamV2Client, error) {
+	stream, err := c.cc.NewStream(ctx, &ECSStateSnapshotService_ServiceDesc.Streams[1], "/ecssnapshot.ECSStateSnapshotService/GetStateLatestStreamV2", opts...)
+	if err != nil {
+		return nil, err
+	}
+	x := &eCSStateSnapshotServiceGetStateLatestStreamV2Client{stream}
+	if err := x.ClientStream.SendMsg(in); err != nil {
+		return nil, err
+	}
+	if err := x.ClientStream.CloseSend(); err != nil {
+		return nil, err
+	}
+	return x, nil
+}
+
+type ECSStateSnapshotService_GetStateLatestStreamV2Client interface {
+	Recv() (*ECSStateReplyV2, error)
+	grpc.ClientStream
+}
+
+type eCSStateSnapshotServiceGetStateLatestStreamV2Client struct {
+	grpc.ClientStream
+}
+
+func (x *eCSStateSnapshotServiceGetStateLatestStreamV2Client) Recv() (*ECSStateReplyV2, error) {
+	m := new(ECSStateReplyV2)
+	if err := x.ClientStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
+}
+
 func (c *eCSStateSnapshotServiceClient) GetStateLatestStreamPruned(ctx context.Context, in *ECSStateRequestLatestStreamPruned, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamPrunedClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ECSStateSnapshotService_ServiceDesc.Streams[1], "/ecssnapshot.ECSStateSnapshotService/GetStateLatestStreamPruned", opts...)
+	stream, err := c.cc.NewStream(ctx, &ECSStateSnapshotService_ServiceDesc.Streams[2], "/ecssnapshot.ECSStateSnapshotService/GetStateLatestStreamPruned", opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -111,6 +151,38 @@ type eCSStateSnapshotServiceGetStateLatestStreamPrunedClient struct {
 
 func (x *eCSStateSnapshotServiceGetStateLatestStreamPrunedClient) Recv() (*ECSStateReply, error) {
 	m := new(ECSStateReply)
+	if err := x.ClientStream.RecvMsg(m); err != nil {
+		return nil, err
+	}
+	return m, nil
+}
+
+func (c *eCSStateSnapshotServiceClient) GetStateLatestStreamPrunedV2(ctx context.Context, in *ECSStateRequestLatestStreamPruned, opts ...grpc.CallOption) (ECSStateSnapshotService_GetStateLatestStreamPrunedV2Client, error) {
+	stream, err := c.cc.NewStream(ctx, &ECSStateSnapshotService_ServiceDesc.Streams[3], "/ecssnapshot.ECSStateSnapshotService/GetStateLatestStreamPrunedV2", opts...)
+	if err != nil {
+		return nil, err
+	}
+	x := &eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Client{stream}
+	if err := x.ClientStream.SendMsg(in); err != nil {
+		return nil, err
+	}
+	if err := x.ClientStream.CloseSend(); err != nil {
+		return nil, err
+	}
+	return x, nil
+}
+
+type ECSStateSnapshotService_GetStateLatestStreamPrunedV2Client interface {
+	Recv() (*ECSStateReplyV2, error)
+	grpc.ClientStream
+}
+
+type eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Client struct {
+	grpc.ClientStream
+}
+
+func (x *eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Client) Recv() (*ECSStateReplyV2, error) {
+	m := new(ECSStateReplyV2)
 	if err := x.ClientStream.RecvMsg(m); err != nil {
 		return nil, err
 	}
@@ -152,8 +224,16 @@ type ECSStateSnapshotServiceServer interface {
 	GetStateLatest(context.Context, *ECSStateRequestLatest) (*ECSStateReply, error)
 	// Requests the latest ECS state in stream format, which will chunk the state.
 	GetStateLatestStream(*ECSStateRequestLatestStream, ECSStateSnapshotService_GetStateLatestStreamServer) error
+	// Requests the latest ECS state in stream format, which will chunk the state.
+	//
+	// V2 version optimized to return entities as raw bytes.
+	GetStateLatestStreamV2(*ECSStateRequestLatestStream, ECSStateSnapshotService_GetStateLatestStreamV2Server) error
 	// Requests the latest ECS state, with aditional pruning.
 	GetStateLatestStreamPruned(*ECSStateRequestLatestStreamPruned, ECSStateSnapshotService_GetStateLatestStreamPrunedServer) error
+	// Requests the latest ECS state, with aditional pruning.
+	//
+	// V2 version optimized to return entities as raw bytes.
+	GetStateLatestStreamPrunedV2(*ECSStateRequestLatestStreamPruned, ECSStateSnapshotService_GetStateLatestStreamPrunedV2Server) error
 	// Requests the latest block number based on the latest ECS state.
 	GetStateBlockLatest(context.Context, *ECSStateBlockRequestLatest) (*ECSStateBlockReply, error)
 	// Requests the ECS state at specific block.
@@ -173,8 +253,14 @@ func (UnimplementedECSStateSnapshotServiceServer) GetStateLatest(context.Context
 func (UnimplementedECSStateSnapshotServiceServer) GetStateLatestStream(*ECSStateRequestLatestStream, ECSStateSnapshotService_GetStateLatestStreamServer) error {
 	return status.Errorf(codes.Unimplemented, "method GetStateLatestStream not implemented")
 }
+func (UnimplementedECSStateSnapshotServiceServer) GetStateLatestStreamV2(*ECSStateRequestLatestStream, ECSStateSnapshotService_GetStateLatestStreamV2Server) error {
+	return status.Errorf(codes.Unimplemented, "method GetStateLatestStreamV2 not implemented")
+}
 func (UnimplementedECSStateSnapshotServiceServer) GetStateLatestStreamPruned(*ECSStateRequestLatestStreamPruned, ECSStateSnapshotService_GetStateLatestStreamPrunedServer) error {
 	return status.Errorf(codes.Unimplemented, "method GetStateLatestStreamPruned not implemented")
+}
+func (UnimplementedECSStateSnapshotServiceServer) GetStateLatestStreamPrunedV2(*ECSStateRequestLatestStreamPruned, ECSStateSnapshotService_GetStateLatestStreamPrunedV2Server) error {
+	return status.Errorf(codes.Unimplemented, "method GetStateLatestStreamPrunedV2 not implemented")
 }
 func (UnimplementedECSStateSnapshotServiceServer) GetStateBlockLatest(context.Context, *ECSStateBlockRequestLatest) (*ECSStateBlockReply, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetStateBlockLatest not implemented")
@@ -238,6 +324,27 @@ func (x *eCSStateSnapshotServiceGetStateLatestStreamServer) Send(m *ECSStateRepl
 	return x.ServerStream.SendMsg(m)
 }
 
+func _ECSStateSnapshotService_GetStateLatestStreamV2_Handler(srv interface{}, stream grpc.ServerStream) error {
+	m := new(ECSStateRequestLatestStream)
+	if err := stream.RecvMsg(m); err != nil {
+		return err
+	}
+	return srv.(ECSStateSnapshotServiceServer).GetStateLatestStreamV2(m, &eCSStateSnapshotServiceGetStateLatestStreamV2Server{stream})
+}
+
+type ECSStateSnapshotService_GetStateLatestStreamV2Server interface {
+	Send(*ECSStateReplyV2) error
+	grpc.ServerStream
+}
+
+type eCSStateSnapshotServiceGetStateLatestStreamV2Server struct {
+	grpc.ServerStream
+}
+
+func (x *eCSStateSnapshotServiceGetStateLatestStreamV2Server) Send(m *ECSStateReplyV2) error {
+	return x.ServerStream.SendMsg(m)
+}
+
 func _ECSStateSnapshotService_GetStateLatestStreamPruned_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(ECSStateRequestLatestStreamPruned)
 	if err := stream.RecvMsg(m); err != nil {
@@ -256,6 +363,27 @@ type eCSStateSnapshotServiceGetStateLatestStreamPrunedServer struct {
 }
 
 func (x *eCSStateSnapshotServiceGetStateLatestStreamPrunedServer) Send(m *ECSStateReply) error {
+	return x.ServerStream.SendMsg(m)
+}
+
+func _ECSStateSnapshotService_GetStateLatestStreamPrunedV2_Handler(srv interface{}, stream grpc.ServerStream) error {
+	m := new(ECSStateRequestLatestStreamPruned)
+	if err := stream.RecvMsg(m); err != nil {
+		return err
+	}
+	return srv.(ECSStateSnapshotServiceServer).GetStateLatestStreamPrunedV2(m, &eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Server{stream})
+}
+
+type ECSStateSnapshotService_GetStateLatestStreamPrunedV2Server interface {
+	Send(*ECSStateReplyV2) error
+	grpc.ServerStream
+}
+
+type eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Server struct {
+	grpc.ServerStream
+}
+
+func (x *eCSStateSnapshotServiceGetStateLatestStreamPrunedV2Server) Send(m *ECSStateReplyV2) error {
 	return x.ServerStream.SendMsg(m)
 }
 
@@ -344,8 +472,18 @@ var ECSStateSnapshotService_ServiceDesc = grpc.ServiceDesc{
 			ServerStreams: true,
 		},
 		{
+			StreamName:    "GetStateLatestStreamV2",
+			Handler:       _ECSStateSnapshotService_GetStateLatestStreamV2_Handler,
+			ServerStreams: true,
+		},
+		{
 			StreamName:    "GetStateLatestStreamPruned",
 			Handler:       _ECSStateSnapshotService_GetStateLatestStreamPruned_Handler,
+			ServerStreams: true,
+		},
+		{
+			StreamName:    "GetStateLatestStreamPrunedV2",
+			Handler:       _ECSStateSnapshotService_GetStateLatestStreamPrunedV2_Handler,
 			ServerStreams: true,
 		},
 	},